refactor: (#129) 중복 room:readyToggle이벤트 통합
