+++ src/game/evaluation/evaluation.gateway.ts
@@ -14,7 +14,6 @@ import {
 } from '../../game-state/interfaces/game-state.interface';
 
 import { EvaluationService } from './evaluation.service';
-import type { RoomUpdatePlayerPayload } from './interfaces/evaluation.interface';
 import { EVALUATION_ERRORS, EVALUATION_EVENTS } from './constants/evaluation.constant';
 import type { GameSocket } from '../interfaces/gameSocket.interface';
 import { SocketExceptionFilter } from 'src/common/filters/socket-exception.filter';
@@ -36,7 +35,6 @@ export class EvaluationGateway {
 
   constructor(
     private readonly evalService: EvaluationService,
-    private readonly gameSessionService: GameSessionService,
     @Inject(GAME_STATE_STORE) private readonly gameStateStore: IGameStateStore,
   ) {}
 
@@ -52,27 +50,5 @@ export class EvaluationGateway {
     if (!gameState) throw wsError(404, EVALUATION_ERRORS.GAME_STATE_NOT_FOUND);
 
     await this.evalService.submitEvaluation(roomId, gameState, userId, payload);
-  }
-
-  @SubscribeMessage(EVALUATION_EVENTS.ROOM_READY_TOGGLE)
-  async onReadyToggle(@ConnectedSocket() client: GameSocket): Promise<void> {
-    const userId = requireUserId(client);
-    const roomId = requireRoomId(client);
-
-    const gameState = await this.gameStateStore.get(roomId);
-    if (!gameState) throw wsError(404, 'GAME_STATE_NOT_FOUND');
-
-    const { isReady, allReady } = await this.evalService.toggleReady(roomId, gameState, userId);
-
-    const payload: RoomUpdatePlayerPayload = {
-      userId,
-      changes: { isReady },
-    };
-
-    this.server.to(`game:room:${roomId}`).emit(EVALUATION_EVENTS.ROOM_UPDATE_PLAYER, payload);
-
-    if (allReady) {
-      await this.gameSessionService.advanceToRoundSummary({ roomId, server: this.server });
-    }
   }
 }
