+++ src/game/interfaces/game.interface.ts
@@ -1,0 +1,6 @@
+export interface RoomUpdatePlayerPayload {
+  userId: number;
+  changes: {
+    isReady: boolean;
+  };
+}
