--- src/game/round-summary/constants/round-summary.constants.ts
@@ -1,4 +1,0 @@
-export const ROUND_SUMMARY_EVENTS = {
-  ROOM_READY_TOGGLE: 'room:readyToggle',
-  ROOM_UPDATE_PLAYER: 'room:updatePlayer',
-} as const;
